Fix timing issue
If `swap` is called at the same time as `access` there is a chance
that the reference counting `Ref` gets updated _after_ the resource
has been finalised.

This introduces an additional semaphore around `access` to ensure
that when the references are incremented for a resource it will
not get accidentally shut down. In practice this means blocking
the creation of the new resource in `swap` while `access` is
invoked.